Factored out fetchRules
